requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Web scraping
scrapy==2.11.0
playwright==1.40.0

# AI/ML
openai==1.3.8

# Payment processing
stripe==7.8.0

# Additional FastAPI dependencies
pydantic==2.5.0

# Rate limiting and security
slowapi==0.1.9
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Database migrations
alembic==1.12.1